LU-13021 wbc: add aging keep cache flush mode for WBC
This flush mode WBC_FLUSH_AGING_KEEP is similar to aging_drop
flush mode. Instead, the WBC EX lock for the root WBC directory
never drops actively unless it has to in the following cases:
- The client needs to revoke the cached root WBC EX lock when the
  directory is conflict accessing from a remote client or shrink
  the LRU locks in the client lock namespace;
- A application or a user wants to cleanup or uncache the cached
  data on the client manually.

This flush mode is the default flush node for WBC.
Before the root WBC EX lock is revoked, nearly all operations for
the files or directories under the root WBC directory can be
executed locally such as open()/close(), getattr(), setattr(),
readdir(), read(), write().
But for the unlink() operation, it should be handled specially
when the files are marked with Sync(s) state. In this patch, it
performs unlink synchronously on MDT for the files in the WBC
state Protected(P) | Sync(S) | Complete(P).

Test-Parameters: clientdistro=el8.3 testlist=sanity-wbc,sanity-wbc
Test-Parameters: testlist=sanity-wbc,sanity-wbc
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I1a6adf41757058b5c6dd5815ee374f8b66ce798d