LU-13010 wbc: reopen files when root WBC EX lock is revoking
For a file flagged with Protected(P) state under the protection of
an EX WBC lock, the open() system call does not need to
communicate with MDS, can just open locally in MemFs on the client
with WBC support.

However, Lustre is a stateful filesystem. Each open keeps certain
state on MDS. It must keep transparent for applications once
deroot an inode once the corresponding root EX WBC lock is
cancelling. To achieve this goal, each local open in MemFS will be
tracked and recorded in a open list per dentry and will be reopned
from MDS when the root WBC EX lock is revoking.

This patch implemented WBC reopen for regular files, all I/O can
directly use the reopened file handle, which is transparent to
applications.

But for directories, it must be handled specially for the
->readdir call.
Currently the mechanism adopted by MemFS (similar to tmpfs/ramfs)
is to simply scan the in-memory sub dentries in dcache linearly
to fill the content for readdir call.
While Lustre new readdir implementation is much complex. It does
readdir in hash order and uses hash of a file name as a telldir/
seekdir cookie stored in the file handle.
Thus, It must bridge the two implementation firstly.

Test-Parameters: clientdistro=el8.3 testlist=sanity-wbc,sanity-wbc
Test-Parameters: testlist=sanity-wbc,sanity-wbc
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I5a40dd440cb7e2c9641a6000e21c5654e5f7e622