LU-15669 wbc: lockless layout write for assimilation
For a batched RPC with creations under WBC, saving the replied
LOVEA into the request buffer for replay means that we may need
to enlarge the request buffer, and this will change the registered
bulk I/O buffer and add the complexity for the replay recovery.

In this patch, we add a option "batch_no_layout". With this
option, the MDT does not create LOVEA for the batched creations,
just creates the metadata object without layout instantiation.

We delay to instantiate the layout until the time of flushing the
cache pages (assimilation phase).
In the assimilation, we implement a lockless reint layout write
to instantiate the data objects for the file.

Add sanity-wbc test_106.

TODO:
1. reply reconstruction for the lockless reint layout write;
2. replay recovery for the lockless reint layout wirte;

Test-Parameters: clientdistro=el8.3 testlist=sanity-wbc,sanity-wbc
Test-Parameters: testlist=sanity-wbc,sanity-wbc
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: Ib6f2fc3e43e452e30bb22d6ed4446d2a29a7c1d0