LU-15799 wbc: add rename support for WBC
The rename() operation uses a strategy similar to hardlink()
operation. It deroots the parent inodes first, and then
synchronizes them to the metadata server to perform the rename()
operation.

Nevertheless, we can do some optimization for rename() operation
under WBC: if the parent directory of the source file and the
target file are both in the state flagged with Complete (C) and
not Flushed (~F), the client can directly perform rename()
operation locally in MemFS.

Test-Parameters: clientdistro=el8.3 testlist=sanity-wbc,sanity-wbc
Test-Parameters: testlist=sanity-wbc,sanity-wbc
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: Iaa0bcb42c73a4431d3177795e9376bd76327fc54