LU-16107 wbc: discard caches when the client is evicted
In the original implementation, the client will flush the subtree
level by level when the root WBC EX lock protected it is revoking.

However, when a client is evicted by a server, it will cancel all
locks cached on the client lock namesapce and all cached data will
be discarded.
During the client eviction, when cancel a WBC EX lock (in the lock
blocking callback), the client will discard the caching for the
whole subtree protected by this lock if the import of the lock is
in EVICTED state.

Add test case sanity-wbc/117, 118 to verify it.

Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I1b148c66ab84bd8b3c12d8a6504ba07c4b60f42a